refactor: extract routes to be resuable in tests
previously, the webserver routes were build in the overly-long
fn main. The route-building is extracted to api.rs to allow the tests
to use this function too.
